requirements.txt
pandas
requests
pendulum
psycopg2-binary
python-dotenv
apache-airflow
airflow-dbt-python
dbt-postgres
apache-airflow-providers-docker